requirements.txt
# List required packages in this file, one per line.
pyyaml
pre-commit
numpy
scipy
pandas
